backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1

# LLM
groq==0.11.0
langchain==0.3.7
langchain-groq==0.2.3
langchain-core==0.3.15
langgraph==0.2.53

# RAG
sentence-transformers==3.1.1
faiss-cpu==1.9.0
rank-bm25==0.2.2
numpy==1.26.4

# Utils
pydantic==2.9.2
